reimplements the customer operations to also put an according representer to the result: an error representer on errors a specific representer on success, refactors the customer controller to use the representers from the operation output directly for forming repsonseenities

+++ src/main/java/training/concepts/customer/operations/Delete.java
@@ -4,6 +4,8 @@ import training.Application;
 import training.concepts.OperationInterface;
 import training.concepts.customer.Customer;
 import training.concepts.customer.CustomerRepository;
+import training.concepts.application.ErrorRepresenter;
+
 import java.util.Map;
 import java.util.HashMap;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,12 +26,24 @@ public class Delete implements OperationInterface {
     Long customerId = (Long) params.get("customerId");
 
     Customer model = repository.findOne(customerId);
-    Application.logger.info(String.format("Found customer: %s", model.toString()));
-    repository.delete(model);
-    Application.logger.info(String.format("Deleted customer!"));
+    Application.logger.info("Found customer: {}", model);
 
-    payload.put("httpStatus", HttpStatus.NO_CONTENT);
-    payload.put("model", null);
-    return payload;
+    if (model == null) {
+      Application.logger.info("Could not find customer with id: {}", customerId);
+      ErrorRepresenter errorRepresenter = new ErrorRepresenter("NOT_FOUND",
+        String.format("The custmer with id: %d could not be found!", customerId));
+      payload.put("httpStatus", HttpStatus.NOT_FOUND);
+      payload.put("representer", errorRepresenter);
+      payload.put("model", null);
+      return payload;
+    } else {
+      repository.delete(model);
+      Application.logger.info(String.format("Deleted customer!"));
+      payload.put("httpStatus", HttpStatus.NO_CONTENT);
+      payload.put("model", null);
+      payload.put("representer", null);
+      return payload;
+    }
+
   }
 }
